0.0.76: Optimize Half Reverse Interpolate Scale (HRIS)
